adk-agent/requirements.txt
# Requirement for running the adk-agent
google-adk==1.14.1
fastapi

agent-starter-pack
psycopg2-binary